use kubernetes in cluster API to query pod list
This removes the need to configure kubelet to enable anonymous authentication and
rely on the standard client-go API of kubernetes.
The authentication now relies on ClusterRole to allow listing pods.

Note that the pods are filtered by nodes as to ensure that only local pods
information are queried. This is done via the downward API:

env:
  - name: NODE_NAME
    valueFrom:
      fieldRef:
        fieldPath: spec.nodeName

Signed-off-by: Gregory Detal <gregory.detal@tessares.net>